// src/components/atoms/button/enums/button.component.enums.ts
enum ButtonComponentAlignmentEnum {
  CENTER = 'center',
  LEFT = 'left',
  RIGHT = 'right',
}

enum ButtonComponentBorderRadiusEnum {
  FULL = 'full',
  LARGE = 'large',
  MEDIUM = 'medium',
  NONE = 'none',
  SMALL = 'small',
}

enum ButtonComponentFontSizeEnum {
  LARGE = 'large',
  MEDIUM = 'medium',
  SMALL = 'small',
}

enum ButtonComponentFontWeightEnum {
  BLACK = 'black',
  BOLD = 'bold',
  EXTRABOLD = 'extrabold',
  EXTRALIGHT = 'extralight',
  LIGHT = 'light',
  MEDIUM = 'medium',
  NORMAL = 'normal',
  SEMIBOLD = 'semibold',
  THIN = 'thin',
}

enum ButtonComponentGapEnum {
  LARGE = 'large',
  MEDIUM = 'medium',
  NONE = 'none',
  SMALL = 'small',
}

enum ButtonComponentModeEnum {
  BUTTON = 'button',
  LINK = 'link',
}

enum ButtonComponentOpacityEnum {
  LOW = 0.25,
  MEDIUM = 0.5,
  HIGH = 0.75,
  FULL = 1.0,
}

enum ButtonComponentPaddingEnum {
  LARGE = 'large',
  MEDIUM = 'medium',
  NONE = 'none',
  SMALL = 'small',
}

enum ButtonComponentShapeEnum {
  CIRCLE = 'circle',
  RECTANGLE = 'rectangle',
  SQUARE = 'square',
}

enum ButtonComponentSizeEnum {
  LARGE = 'large',
  MEDIUM = 'medium',
  SMALL = 'small',
}

enum ButtonComponentVariantEnum {
  DESTRUCTIVE = 'destructive',
  GHOST = 'ghost',
  HIDDEN = 'hidden',
  INVISIBLE = 'invisible',
  OUTLINE = 'outline',
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
}

export {
  ButtonComponentAlignmentEnum,
  ButtonComponentBorderRadiusEnum,
  ButtonComponentFontSizeEnum,
  ButtonComponentFontWeightEnum,
  ButtonComponentGapEnum,
  ButtonComponentModeEnum,
  ButtonComponentOpacityEnum,
  ButtonComponentPaddingEnum,
  ButtonComponentShapeEnum,
  ButtonComponentSizeEnum,
  ButtonComponentVariantEnum,
};